+++ moPepGen/svgraph/ThreeFrameTVG.py
@@ -1702,9 +1702,22 @@ class ThreeFrameTVG():
 
                 # create new node with the combined sequence
                 new_node = cur.copy()
-                was_bridge = new_node.reading_frame_index != rf_index \
+                # out_node.was_bridge is added for the case the the out node is
+                # processed first, and is merged from nodes carrying multiple
+                # frameshift variants which together become in-frame.
+                was_bridge = out_node.was_bridge or (
+                    new_node.reading_frame_index != rf_index
                     and any(x.reading_frame_index != rf_index for x in out_node.get_out_nodes())
+                )
+
                 new_node.append_right(out_node)
+                # This is a better way to tell whether it was a bridge node..
+                # new_node.was_bridge = (
+                #     sum(v.variant.frames_shifted() for v in cur.variants) % 3 > 0
+                #     or cur.was_bridge
+                #     or sum(v.variant.frames_shifted() for v in out_node.variants) % 3 > 0
+                #     or out_node.was_bridge
+                # )
                 new_node.was_bridge = was_bridge
                 if new_node.level < out_node.level:
                     new_node.subgraph_id = out_node.subgraph_id
@@ -1901,6 +1914,8 @@ class ThreeFrameTVG():
                     cur = self.merge_with_outbonds(cur)[0]
                 queue.appendleft(cur)
                 continue
+
+            cur_seq = str(cur.seq.seq)
 
             try:
                 self.align_variants(cur)