+++ CHANGELOG.md
@@ -32,6 +32,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Updated `fuzzTest` to pass codon tabel and start codons to `callVariant` and `bruteForce`
 
+- Fixed `callVariant` that during TVG alignment, node merged from multiple frameshifts which together go back to the origianl frame, was not recognized as `was_brige` in a bubble
+
 ## [1.4.6] - 2025-05-21
 
 - Fixed biopython version #908